+++ hnn_core/dipole.py
@@ -95,7 +95,7 @@ def simulate_dipole(net, n_trials=None, record_vsoma=False,
     net._instantiate_drives(n_trials=n_trials)
     net.cell_response.reset()  # see #290 for context; relevant for MPI
     for arr in net.rec_array.values():
-        arr.reset()  # overwrite, not append, any previous simulation results
+        arr._reset()  # overwrite, not append, any previous simulation results
 
     if isinstance(record_vsoma, bool):
         net._params['record_vsoma'] = record_vsoma